+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/CosmosQueryRequestOptionsBase.java
@@ -44,6 +44,7 @@ public abstract class CosmosQueryRequestOptionsBase<T extends CosmosQueryRequest
     private CosmosDiagnosticsThresholds thresholds;
     private Map<String, String> customOptions;
     private boolean indexMetricsEnabled;
+    private boolean queryAdviceEnabled;
     private UUID correlationActivityId;
     private CosmosEndToEndOperationLatencyPolicyConfig cosmosEndToEndOperationLatencyPolicyConfig;
     private List<String> excludeRegions;
@@ -76,6 +77,7 @@ public abstract class CosmosQueryRequestOptionsBase<T extends CosmosQueryRequest
         this.dedicatedGatewayRequestOptions = options.dedicatedGatewayRequestOptions;
         this.customOptions = options.customOptions;
         this.indexMetricsEnabled = options.indexMetricsEnabled;
+        this.queryAdviceEnabled = options.queryAdviceEnabled;
         this.correlationActivityId = options.correlationActivityId;
         this.thresholds = options.thresholds;
         this.cosmosEndToEndOperationLatencyPolicyConfig = options.cosmosEndToEndOperationLatencyPolicyConfig;
@@ -440,6 +442,34 @@ public abstract class CosmosQueryRequestOptionsBase<T extends CosmosQueryRequest
     }
 
     /**
+     * Gets queryAdviceEnabled, which is used to obtain query advice to understand aspects of the query
+     * that can be optimized. The results will be displayed in FeedResponse.getQueryAdvice().
+     * Please note that this option will incur additional latency overhead, so it should be enabled
+     * when debugging queries.
+     *
+     * @return queryAdviceEnabled (default: false)
+     */
+    @Override
+    public Boolean isQueryAdviceEnabled() {
+        return queryAdviceEnabled;
+    }
+
+    /**
+     * Sets queryAdviceEnabled, which is used to obtain query advice to understand aspects of the query
+     * that can be optimized. The results will be displayed in FeedResponse.getQueryAdvice().
+     * Please note that this option will incur additional latency overhead, so it should be enabled
+     * when debugging queries. By default, query advice is disabled.
+     *
+     * @param queryAdviceEnabled a boolean used to obtain the query advice
+     * @return the current request options instance
+     */
+    @SuppressWarnings("unchecked")
+    public T setQueryAdviceEnabled(boolean queryAdviceEnabled) {
+        this.queryAdviceEnabled = queryAdviceEnabled;
+        return (T)this;
+    }
+
+    /**
      * Sets the custom query request option value by key
      *
      * @param name  a string representing the custom option's name
@@ -520,6 +550,7 @@ public abstract class CosmosQueryRequestOptionsBase<T extends CosmosQueryRequest
         this.excludeRegions = overrideOption(cosmosRequestOptions.getExcludedRegions(), this.excludeRegions);
         this.thresholds = overrideOption(cosmosRequestOptions.getDiagnosticsThresholds(), this.thresholds);
         this.indexMetricsEnabled = overrideOption(cosmosRequestOptions.isIndexMetricsEnabled(), this.indexMetricsEnabled);
+        this.queryAdviceEnabled = overrideOption(cosmosRequestOptions.isQueryAdviceEnabled(), this.queryAdviceEnabled);
         this.queryMetricsEnabled = overrideOption(cosmosRequestOptions.isQueryMetricsEnabled(), this.queryMetricsEnabled);
         this.responseContinuationTokenLimitInKb = overrideOption(cosmosRequestOptions.getResponseContinuationTokenLimitInKb(), this.responseContinuationTokenLimitInKb);
         this.keywordIdentifiers = overrideOption(cosmosRequestOptions.getKeywordIdentifiers(), this.keywordIdentifiers);