+++ sdk/dataprotection/azure-resourcemanager-dataprotection/src/main/java/com/azure/resourcemanager/dataprotection/implementation/ResourceGuardsImpl.java
@@ -34,12 +34,8 @@ public final class ResourceGuardsImpl implements ResourceGuards {
         Response<DppBaseResourceInner> inner = this.serviceClient()
             .getDefaultDeleteResourceGuardProxyRequestsObjectWithResponse(resourceGroupName, resourceGuardsName,
                 requestName, context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new DppBaseResourceImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
+        return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
+            new DppBaseResourceImpl(inner.getValue(), this.manager()));
     }
 
     public DppBaseResource getDefaultDeleteResourceGuardProxyRequestsObject(String resourceGroupName,
@@ -71,12 +67,8 @@ public final class ResourceGuardsImpl implements ResourceGuards {
         String resourceGuardsName, Context context) {
         Response<ResourceGuardResourceInner> inner
             = this.serviceClient().getByResourceGroupWithResponse(resourceGroupName, resourceGuardsName, context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new ResourceGuardResourceImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
+        return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
+            new ResourceGuardResourceImpl(inner.getValue(), this.manager()));
     }
 
     public ResourceGuardResource getByResourceGroup(String resourceGroupName, String resourceGuardsName) {
@@ -124,12 +116,8 @@ public final class ResourceGuardsImpl implements ResourceGuards {
         Response<DppBaseResourceInner> inner = this.serviceClient()
             .getDefaultDisableSoftDeleteRequestsObjectWithResponse(resourceGroupName, resourceGuardsName, requestName,
                 context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new DppBaseResourceImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
+        return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
+            new DppBaseResourceImpl(inner.getValue(), this.manager()));
     }
 
     public DppBaseResource getDefaultDisableSoftDeleteRequestsObject(String resourceGroupName,
@@ -162,12 +150,8 @@ public final class ResourceGuardsImpl implements ResourceGuards {
         Response<DppBaseResourceInner> inner = this.serviceClient()
             .getDefaultUpdateProtectedItemRequestsObjectWithResponse(resourceGroupName, resourceGuardsName, requestName,
                 context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new DppBaseResourceImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
+        return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
+            new DppBaseResourceImpl(inner.getValue(), this.manager()));
     }
 
     public DppBaseResource getDefaultUpdateProtectedItemRequestsObject(String resourceGroupName,
@@ -200,12 +184,8 @@ public final class ResourceGuardsImpl implements ResourceGuards {
         Response<DppBaseResourceInner> inner = this.serviceClient()
             .getDefaultUpdateProtectionPolicyRequestsObjectWithResponse(resourceGroupName, resourceGuardsName,
                 requestName, context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new DppBaseResourceImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
+        return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
+            new DppBaseResourceImpl(inner.getValue(), this.manager()));
     }
 
     public DppBaseResource getDefaultUpdateProtectionPolicyRequestsObject(String resourceGroupName,
@@ -238,12 +218,8 @@ public final class ResourceGuardsImpl implements ResourceGuards {
         Response<DppBaseResourceInner> inner = this.serviceClient()
             .getDefaultDeleteProtectedItemRequestsObjectWithResponse(resourceGroupName, resourceGuardsName, requestName,
                 context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new DppBaseResourceImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
+        return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
+            new DppBaseResourceImpl(inner.getValue(), this.manager()));
     }
 
     public DppBaseResource getDefaultDeleteProtectedItemRequestsObject(String resourceGroupName,
@@ -276,12 +252,8 @@ public final class ResourceGuardsImpl implements ResourceGuards {
         Response<DppBaseResourceInner> inner = this.serviceClient()
             .getDefaultBackupSecurityPinRequestsObjectWithResponse(resourceGroupName, resourceGuardsName, requestName,
                 context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new DppBaseResourceImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
+        return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
+            new DppBaseResourceImpl(inner.getValue(), this.manager()));
     }
 
     public DppBaseResource getDefaultBackupSecurityPinRequestsObject(String resourceGroupName,