+++ sdk/recoveryservicesbackup/azure-resourcemanager-recoveryservicesbackup/src/main/java/com/azure/resourcemanager/recoveryservicesbackup/implementation/OperationOperationsImpl.java
@@ -31,12 +31,8 @@ public final class OperationOperationsImpl implements OperationOperations {
         ValidateOperationRequestResource parameters, Context context) {
         Response<ValidateOperationsResponseInner> inner
             = this.serviceClient().validateWithResponse(vaultName, resourceGroupName, parameters, context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new ValidateOperationsResponseImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
+        return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
+            new ValidateOperationsResponseImpl(inner.getValue(), this.manager()));
     }
 
     public ValidateOperationsResponse validate(String vaultName, String resourceGroupName,