+++ sdk/recoveryservicesbackup/azure-resourcemanager-recoveryservicesbackup/src/main/java/com/azure/resourcemanager/recoveryservicesbackup/implementation/PrivateEndpointsImpl.java
@@ -31,12 +31,8 @@ public final class PrivateEndpointsImpl implements PrivateEndpoints {
         Response<OperationStatusInner> inner = this.serviceClient()
             .getOperationStatusWithResponse(vaultName, resourceGroupName, privateEndpointConnectionName, operationId,
                 context);
-        if (inner != null) {
-            return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
-                new OperationStatusImpl(inner.getValue(), this.manager()));
-        } else {
-            return null;
-        }
+        return new SimpleResponse<>(inner.getRequest(), inner.getStatusCode(), inner.getHeaders(),
+            new OperationStatusImpl(inner.getValue(), this.manager()));
     }
 
     public OperationStatus getOperationStatus(String vaultName, String resourceGroupName,