+++ sdk/cosmos/azure-cosmos-spark_3-3_2-12/CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Release History
 
+### 4.44.2 (2026-03-05)
+
+#### Other Changes
+* Changed azure-resourcemanager-cosmos usage to a pinned version which is deployed across all public and non-public clouds - [PR 48268](https://github.com/Azure/azure-sdk-for-java/pull/48268)
+
+### 4.44.1 (2026-03-03)
+
+#### Other Changes
+* Reduced noisy warning logs in Gateway mode - [PR 48189](https://github.com/Azure/azure-sdk-for-java/pull/48189)
+
 ### 4.44.0 (2026-02-27)
 
 #### Features Added